déplacement hsami_inputs et renomer main
